fix: correctly show validations messages on invalid input
