+++ frontend/static/js/main.js
@@ -1,10 +1,10 @@
 import authPage from '../pages/auth.html?raw'
 import homePage from '../pages/home.html?raw'
-import { settings } from './2FA.js'
+import { twoFA } from './2FA.js'
 import { login, login42, logout } from './auth.js'
 import { chat } from './chat.js'
 import Client from './client.js'
-import { editProfile, updateProfile } from './edit-profile.js'
+import { editProfile } from './edit-profile.js'
 import { friends } from './friends.js'
 import { updateNavbar } from './navbar.js'
 import { choosePongMode, joinGame, localGame, remotePong, remoteSetup } from './pong-setup.js'
@@ -48,6 +48,5 @@ router.get('/pong/remote/join/:id', (client, params) => joinGame(client, params.
 router.get('/pong/local', localGame)
 
 router.get('/profile/edit', editProfile)
-router.get('/profile/update', updateProfile)
 router.get('/friends', friends)
-router.get('/settings', settings)
+router.get('/profile/2FA', twoFA)
