+++ frontend/static/pages/edit-profile.html
@@ -1,30 +1,32 @@
 <div class="container-edit-profile">
   <div id="edit-card" class="card mb-4">
+    <div id="edit-alert" class="alert alert-danger d-none" role="alert">Please fill all the fields.</div>
     <div class="edit-content">
       <div class="edit-info">
          <h1 class="mb-4">Edit Profile</h1>
         <div class="mb-3-edit-profile text-start">
           <label for="username"><strong>Username:</strong></label>
-          <input type="text" id="username"><br>
-          <div class="invalid-feedback" id="username-feedback">This username is already used.</div>
+          <input type="text" id="username">
+          <div id="username-empty-feedback" class="invalid-feedback d-none" >Please enter a username.</div>
+          <div id="username-exists-feedback" class="invalid-feedback d-none">This username is already used.</div>
         </div>
         <div class="mb-3-edit-profile text-start">
           <label for="email"><strong>Email:</strong></label>
-          <input type="email" id="email" class="edit-control"><br>
-          <div class="invalid-feedback" id="email-format-feedback">Please enter a valid email address.</div>
-          <div class="invalid-feedback" id="email-exists-feedback" style="display: none;">This email is already used.</div>
+          <input type="email" id="email" class="edit-control">
+          <div id="email-format-feedback" class="invalid-feedback d-none" >Please enter a valid email address.</div>
+          <div id="email-exists-feedback" class="invalid-feedback d-none" >This email is already used.</div>
         </div>
         <div class="mb-3-edit-profile text-start">
           <label for="nickname"><strong>Nickname:</strong></label>
-          <input type="text" class="edit-control" id="nickname"><br>
-          <div id="nickname-exists-feedback" class="invalid-feedback">This nickname is already used.</div>
+          <input type="text" class="edit-control" id="nickname">
+          <div id="nickname-empty-feedback" class="invalid-feedback d-none">Please enter a nickname.</div>
+          <div id="nickname-exists-feedback" class="invalid-feedback d-none">This nickname is already used.</div>
         </div>
          <div class="mb-3-edit-profile text-start">
         <label for="new-avatar"><strong>Avatar:</strong></label>
-        <input type="file" id="new-avatar" class="edit-control" accept="image/*"><br>
+        <input type="file" id="new-avatar" class="edit-control" accept="image/*">
         </div>
-        <div id="edit-alert" class="alert alert-danger d-none" role="alert">Please fill all the fields</div>
-        <a href="/profile/update" id="save-profile-info-link" class="btn btn-primary">Save Profile Info</a>
+        <button id="save-profile-info-link" class="btn btn-primary">Save Profile Info</button>
       </div>
     </div>
   </div>
